[Time] 05-02-2019 | 16:00
[Description] Floating Shortcuts | Quickly Switch Between Apps by Floating Shortcuts | Multitasking Master 📱⌚️

[Phone][ChangeLog]
~ Fast Scrolling Indexing with Popup Indicator
- Hybrid List Apps
- Folders Apps Selection
- Configured Widgets List
- Installed Widgets List

[Watch][ChangeLog]
~

[Build][ChangeLog]
~

+++ FloatingShortcutsPhone/src/main/java/net/geekstools/floatshort/PRO/Category/AppSelectionList.java
@@ -69,7 +69,7 @@ public class AppSelectionList extends Activity implements View.OnClickListener {
 
     List<ApplicationInfo> applicationInfoList;
     Map<String, Integer> mapIndexFirstItem, mapIndexLastItem;
-    Map<Float, String> mapRangeIndex;
+    Map<Integer, String> mapRangeIndex;
     ArrayList<NavDrawerItem> navDrawerItems, navDrawerItemsSaved;
     RecyclerView.Adapter appSelectionListAdapter;
     LinearLayoutManager recyclerViewLayoutManager;
@@ -124,7 +124,7 @@ public class AppSelectionList extends Activity implements View.OnClickListener {
         navDrawerItemsSaved = new ArrayList<NavDrawerItem>();
         mapIndexFirstItem = new LinkedHashMap<String, Integer>();
         mapIndexLastItem = new LinkedHashMap<String, Integer>();
-        mapRangeIndex = new LinkedHashMap<Float, String>();
+        mapRangeIndex = new LinkedHashMap<Integer, String>();
 
         Typeface face = Typeface.createFromAsset(getAssets(), "upcil.ttf");
         desc.setTypeface(face);
@@ -533,11 +533,11 @@ public class AppSelectionList extends Activity implements View.OnClickListener {
             new Handler().postDelayed(new Runnable() {
                 @Override
                 public void run() {
-                    float upperRange = indexView.getY() - finalTextView.getHeight();
+                    int upperRange = (int) (indexView.getY() - finalTextView.getHeight());
                     for (int i = 0; i < indexView.getChildCount(); i++) {
                         String indexText = ((TextView) indexView.getChildAt(i)).getText().toString();
-                        float indexRange = (indexView.getChildAt(i).getY() + indexView.getY() + finalTextView.getHeight());
-                        for (float jRange = upperRange; jRange <= (indexRange); jRange++) {
+                        int indexRange = (int) (indexView.getChildAt(i).getY() + indexView.getY() + finalTextView.getHeight());
+                        for (int jRange = upperRange; jRange <= (indexRange); jRange++) {
                             mapRangeIndex.put(jRange, indexText);
                         }
 
@@ -565,7 +565,7 @@ public class AppSelectionList extends Activity implements View.OnClickListener {
             public boolean onTouch(View view, MotionEvent motionEvent) {
                 switch (motionEvent.getAction()) {
                     case MotionEvent.ACTION_DOWN: {
-                        String indexText = mapRangeIndex.get(motionEvent.getY());
+                        String indexText = mapRangeIndex.get(((int) motionEvent.getY()));
 
                         if (indexText != null) {
                             popupIndex.setY(motionEvent.getRawY() - popupIndexOffsetY);
@@ -577,7 +577,7 @@ public class AppSelectionList extends Activity implements View.OnClickListener {
                         break;
                     }
                     case MotionEvent.ACTION_MOVE: {
-                        String indexText = mapRangeIndex.get(motionEvent.getY());
+                        String indexText = mapRangeIndex.get(((int) motionEvent.getY()));
 
                         if (indexText != null) {
                             if (!popupIndex.isShown()) {
@@ -590,7 +590,7 @@ public class AppSelectionList extends Activity implements View.OnClickListener {
                             try {
                                 nestedScrollView.smoothScrollTo(
                                         0,
-                                        ((int) loadView.getChildAt(mapIndexFirstItem.get(mapRangeIndex.get(motionEvent.getY()))).getY())
+                                        ((int) loadView.getChildAt(mapIndexFirstItem.get(mapRangeIndex.get(((int) motionEvent.getY())))).getY())
                                 );
                             } catch (Exception e) {
                                 e.printStackTrace();
@@ -609,7 +609,7 @@ public class AppSelectionList extends Activity implements View.OnClickListener {
                             try {
                                 nestedScrollView.smoothScrollTo(
                                         0,
-                                        ((int) loadView.getChildAt(mapIndexFirstItem.get(mapRangeIndex.get(motionEvent.getY()))).getY())
+                                        ((int) loadView.getChildAt(mapIndexFirstItem.get(mapRangeIndex.get(((int) motionEvent.getY())))).getY())
                                 );
                             } catch (Exception e) {
                                 e.printStackTrace();
